Added RealTimeMessaging
An RTM session is started by firs issuing rtm_start. This sends the
command over to the Slack server, which then returns an URL you are
expected to connect via websocket. You provide this url to rtm_connect
which returns an rtm_session, which is your handle to the connection.

With rtm_session you can do Slacko.rtm_receive to receive RTM
events. So far only private messages are supported. Other messages are
dropped.

The session automatically sends pings every 5 seconds if there is no
other activity. In addition, if the connection doesn't give any data
back (ie. ping responses) for 20 seconds, it is disconnected. In that
case (and in other cases when the connection is disconneceed)
rtm_receive returns with RtmDisconnected.

To disconnect the session you can call rtm_stop.